// generated/src/app/api-models/member-contact.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberContact {
  seqMembId: number;
  contactSource: string;
  primaryDistributionMethod: string;
  emailId: string;
  faxNumber: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}